backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart # For OAuth2PasswordRequestForm
sqlalchemy
psycopg2-binary # For PostgreSQL connection
alembic # For database migrations (optional for now, but good practice)
python-dotenv # For loading .env files
redis # For Redis cache integration
# requests # Might be needed if backend calls AI agent directly, or for other external services
